[Refactor] Separated View Bills Component
