+++ frontend/src/pages/NotFound.jsx
@@ -1,8 +1,10 @@
 const NotFound = () => {
-    return <div>
-        <h1>404 Not Found</h1>
-        <p>The page you're looking for does not exist!</p>
-    </div>
-}
+  return (
+    <>
+      <h1>404 Not Found</h1>
+      <p>The page you're looking for does not exist!</p>
+    </>
+  );
+};
 
 export default NotFound;
